Crate zaplib_components

Collection of widgets for use with Zaplib.

Doesn’t contain lower level primitives; those are in zaplib.

Structs

AnimFolding

AnimSelect

ArrowPointerIns

Background

Button

Chart

ChartConfig

These options are based on the ones provided by ChartJS

ChartCurrentElement

ChartDataset

ChartScale

ChartStyle

ChartTooltip

ChartTooltipConfig

Tooltip configuration

Checkbox

CodeEditorColors

ColorBackground

Cursor

CursorIns

CursorRect

DesktopWindow

Dock

DockTab

DockTabIdent

DockWalker

DrawCursors

DrawLines3d

DrawLines3dInstance

DrawLines3dOptions

Renderer for line markers.

DrawPoints3d

DrawPoints3dInstance

DrawPoints3dOptions

DrawSel

FloatSlider

FloatSliderBackgroundRange

FoldCaption

FpsCounter

Geometry3d

Represents a single vertex used in 3d objects.

IndentLines

LineGeom

LineTokenizer

List

ListItem

ParenItem

Popover

Popover that can be used for menus, tooltips, etc.

ScrollBarConfig

ScrollShadow

ScrollView

SelectScroll

SelectionIns

SkyBox

SphericalAngles

Spherical coordinates follow the same conventions as https://threejs.org/docs/#api/en/math/Spherical

Splitter

Tab

TabControl

TextBuffer

TextBufferKeyboard

TextBufferMarkers

TextBufferMessage

TextCursor

TextCursorSet

TextEditor

TextInput

TextInputOptions

TextInputTokenizer

TextOp

TextPos

TextUndo

TokenChunk

TokenizerState

Viewport3D

Viewport3DProps

Enums

AnimFoldingState

ArrowPointerDirection

Axis

ButtonEvent

ChartData

Contains the data is going to be used to render the chart

ChartEvent

ChartType

TODO(hernan): Implement other chart types like bar, pie, etc…

CheckboxEvent

Coordinates

DesktopWindowEvent

DockDragEnd

DockEvent

DockItem

DrawPoints3dStyle

FloatSliderEvent

FoldOpenState

ListEvent

ListLogicEvent

ListSelect

SplitterAlign

SplitterEvent

TabControlEvent

TabEvent

TextBufferMessageLevel

TextEditorEvent

TextUndoGrouping

TokenType

Constants

CHART_STYLE_DARK

CHART_STYLE_LIGHT

Traits

ChartPlugin

An extension machanism for charts

ChartTooltipRenderer

Renders a tooltip’s content