Trait zaplib_components::ChartPlugin

pub trait ChartPlugin {
    fn draw(
        &mut self, 
        cx: &mut Cx, 
        config: &ChartConfig<'_>, 
        chart_bounds: &Rect
    );
}

An extension machanism for charts

Implement this trait to render custom elements on top of the chart, but behind tooltips, if any.

Note that any new element added by this function will not be considered for selection and will not affect selection of existing elements drawn by the chart.

Required methods

fn draw(&mut self, cx: &mut Cx, config: &ChartConfig<'_>, chart_bounds: &Rect)